优化下载数据

+++ download/industry.go
@@ -39,7 +39,7 @@ func UpdateMongo(items []map[string]interface{}) {
 	myFunc := func(s []map[string]interface{}) {
 		myBulk := RealColl.Bulk()
 
-		// 初始化事务
+		// 初始化事务：写入数据
 		for i := range s {
 			myBulk = myBulk.UpdateId(s[i]["code"], bson.M{"$set": s[i]})
 		}
@@ -101,7 +101,7 @@ func CalIndustry() {
 				continue
 			}
 			i["tr"] = float64(i["vol"].(int32)) / i["float_share"].(float64) * 10000
-			i["pct_chg"] = i["power"].(float64) / i["fmc"].(float64)
+			i["pct_chg"] = i["power"].(float64) / float64(i["fmc"].(int64))
 			i["type"] = idsName
 			i["marketType"] = "CN"
 			if idsName == "sw" {
